// shiku-world-medium/client/communication/api/bindings/ShowEntity.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { EntityRenderData } from "./EntityRenderData";

export interface ShowEntity { id: string, initial_isometrics_2d: [number, number, number], render: EntityRenderData, parent_entity: string | null, }