// shiku-world-medium/client/communication/api/bindings/LayerName.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

export type LayerName = "Terrain" | "Empty" | "GameObjects" | "Guest" | "BG0" | "BG1" | "BG2" | "BG3" | "BG4" | "BG5" | "BG6" | "BG7" | "BG8" | "BG9" | "BG10" | "BG11" | "FG0" | "FG1" | "FG2" | "FG3" | "FG4" | "FG5" | "FG6" | "FG7" | "FG8" | "FG9" | "FG10" | "FG11" | "Menu";