// shiku-world-medium/client/communication/api/bindings/StaticImage.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { LayerName } from "./LayerName";

export interface StaticImage { graphic_id: string, layer: LayerName, offset_2d: [number, number], scale: [number, number], width: number | null, height: number | null, tiled: boolean, blending_mode: number | null, }