// shiku-world-medium/client/communication/api/bindings/SimpleImageEffect.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { LayerName } from "./LayerName";

export interface SimpleImageEffect { initial_isometrics_2d: [number, number, number], graphic_id: string, blending_mode: number | null, transparency: number, layer: LayerName, parent_entity: string | null, }